+++ src/Traits/LogsActivity.php
@@ -5,7 +5,6 @@ namespace Spatie\Activitylog\Traits;
 use Illuminate\Support\Collection;
 use Spatie\Activitylog\ActivityLogger;
 use Illuminate\Database\Eloquent\Model;
-use Spatie\Activitylog\Models\Activity;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Spatie\Activitylog\ActivitylogServiceProvider;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
@@ -13,6 +12,8 @@ use Illuminate\Database\Eloquent\Relations\MorphMany;
 trait LogsActivity
 {
     use DetectsChanges;
+
+    protected $enableLoggingModelsEvents = true;
 
     protected static function bootLogsActivity()
     {
@@ -37,6 +38,20 @@ trait LogsActivity
                     ->log($description);
             });
         });
+    }
+
+    public function disableLogging()
+    {
+        $this->enableLoggingModelsEvents = false;
+
+        return $this;
+    }
+
+    public function enableLogging()
+    {
+        $this->enableLoggingModelsEvents = true;
+
+        return $this;
     }
 
     public function activity(): MorphMany
@@ -69,7 +84,7 @@ trait LogsActivity
             'deleted',
         ]);
 
-        if (collect(class_uses(__CLASS__))->contains(SoftDeletes::class)) {
+        if (collect(class_uses_recursive(__CLASS__))->contains(SoftDeletes::class)) {
             $events->push('restored');
         }
 
@@ -87,6 +102,10 @@ trait LogsActivity
 
     protected function shouldLogEvent(string $eventName): bool
     {
+        if (! $this->enableLoggingModelsEvents) {
+            return false;
+        }
+
         if (! in_array($eventName, ['created', 'updated'])) {
             return true;
         }